Added shadows for box, added nav buttons for forgot pass and registration
