+++ src/components/Auth/Auth.scss
@@ -11,6 +11,7 @@
     width: 50%;
     height: 50%;
     border-radius: 1rem;
+    box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
 
     form {
       align-items: center;
@@ -24,6 +25,17 @@
       .auth-error {
         color: red;
         text-align: center;
+      }
+
+      a {
+        font-size: .8rem;
+        text-decoration: none;
+        color: #ccc;
+        transition: .4s;
+      }
+
+      a:hover, a.active {
+        color: #000
       }
     }
   }
