+++ src/components/Auth/ForgotPassword.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const ForgotPassword = () => {
+
+  return (
+    <>
+    </>
+  )
+}
+
+export default ForgotPassword
